# backend/tests/fixtures/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AgeVerificationRequestResponse": {
        "properties": {
          "accountId": {
            "format": "uuid",
            "title": "Accountid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reviewedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewedat"
          },
          "reviewedBy": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewedby"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "accountId",
          "status",
          "createdAt"
        ],
        "title": "AgeVerificationRequestResponse",
        "type": "object"
      },
...
      "Body_submit_api_me_age_verification_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_submit_api_me_age_verification_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReasonRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "ReasonRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "\ud55c\uc904 ebook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/age-verification": {
      "get": {
        "description": "\ud604\uc7ac \uc9c4\ud589\uc911(PENDING) \uc778\uc99d \uc694\uccad \u2014 \uc5c6\uc73c\uba74 null(\uacfc\uac70 \uc2b9\uc778/\uac70\ubd80 \uc774\ub825\uc740 \ub178\ucd9c \uc548 \ud568).",
        "operationId": "my_status_api_me_age_verification_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AgeVerificationRequestResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response My Status Api Me Age Verification Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Status",
        "tags": [
          "age-verification"
        ]
      },
      "post": {
        "description": "\uc2e0\ubd84\uc99d \uc0ac\uc9c4 \uc5c5\ub85c\ub4dc \u2192 \uc131\uc778\uc778\uc99d \uc2ec\uc0ac \uc694\uccad. AlreadyPending(409) \u2192 \uc911\uc559 \ud578\ub4e4\ub7ec.",
        "operationId": "submit_api_me_age_verification_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_api_me_age_verification_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgeVerificationRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit",
        "tags": [
          "age-verification"
        ]
      }
    },
...
    "/api/potato/age-verification": {
      "get": {
        "operationId": "list_pending_api_potato_age_verification_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "PENDING",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgeVerificationRequestResponse"
                  },
                  "title": "Response List Pending Api Potato Age Verification Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pending",
        "tags": [
          "potato"
        ]
      }
    },
    "/api/potato/age-verification/{request_id}/approve": {
      "post": {
        "description": "\uc2b9\uc778 \u2014 verified_tier_cd=AGE18 \uac31\uc2e0 + \uc6d0\ubcf8 \uc774\ubbf8\uc9c0 \uc0ad\uc81c. InvalidRequestState(409) \u2192 \uc911\uc559 \ud578\ub4e4\ub7ec.",
        "operationId": "approve_api_potato_age_verification__request_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "potato"
        ]
      }
    },
    "/api/potato/age-verification/{request_id}/photo": {
      "get": {
        "description": "\uc2ec\uc0ac\uc6a9 \uc2e0\ubd84\uc99d \uc6d0\ubcf8 \uc870\ud68c \u2014 RequestNotFound(404, \uc911\uc559 \ud578\ub4e4\ub7ec)\uba74 \ubbf8\uc874\uc7ac/\uc774\ubbf8 \uc0ad\uc81c\ub428.",
        "operationId": "get_photo_api_potato_age_verification__request_id__photo_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Photo",
        "tags": [
          "potato"
        ]
      }
    },
    "/api/potato/age-verification/{request_id}/reject": {
      "post": {
        "description": "\uac70\ubd80 \u2014 \uc6d0\ubcf8 \uc774\ubbf8\uc9c0 \uc0ad\uc81c. \uc0ac\uc720\ub294 \uac10\uc0ac\ub85c\uadf8\uc5d0\ub9cc \ub0a8\uae40(\uc694\uccad \ud14c\uc774\ube14\uc5d4 \uc800\uc7a5\ud558\uc9c0 \uc54a\uc74c).",
        "operationId": "reject_api_potato_age_verification__request_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject",
        "tags": [
          "potato"
        ]
      }
    },
...
  }
}